# server/src/main/resources/openapi/functions.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowCatalyst Function Platform API",
    "version": "1.0.0",
    "description": "The function-platform surface: `/api/functions…`, `/api/function-pools`, `/api/function-policies…`, `/api/function-domains…`, `/api/function-routes`, and the function-host control plane under `/control/functions/…`. Hand-authored (spec `docs/spec/function-openapi.md`) — Java-first, outside `openapi.lock.json` (there is no Go for this surface). NOT here: the invocation surface itself, `/functions/{address}[:{version}]/{path}` on a function host (a passthrough to the function's own endpoints, `docs/spec/function-invocation.md`), and the platform's main API, served at `/api/openapi.json`."
  },
...
  "paths": {
...
    "/api/function-domains": {
      "post": {
        "operationId": "claimFunctionDomain",
        "tags": [
          "function-domains"
        ],
        "summary": "Claim a domain (zone)",
        "description": "Claims `hostname` and every hostname under it — a zone, not a single exact hostname. Requires `platform:function:domain:manage`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The claimed domain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "HOSTNAME_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DOMAIN_TAKEN — an existing claim (any owner) equals, covers, or is covered by this hostname",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "listFunctionDomains",
        "tags": [
          "function-domains"
        ],
        "summary": "List domains owned by a client (or `platform`)",
        "description": "Requires `platform:function:function:view`.",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "required": true,
            "description": "an owning client id, or `platform`",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DomainResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "CLIENT_ID_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/function-domains/{hostname}/verify": {
      "post": {
        "operationId": "verifyFunctionDomain",
        "tags": [
          "function-domains"
        ],
        "summary": "Verify a claimed zone's TXT record",
        "description": "Requires `platform:function:domain:manage`.",
        "parameters": [
          {
            "name": "hostname",
            "in": "path",
            "required": true,
            "description": "any hostname covered by the claimed zone, or the zone apex itself — resolves to the zone's one claim",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The verified domain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "FunctionDomain_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DOMAIN_NOT_VERIFIED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DNS_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/function-domains/{hostname}": {
      "get": {
        "operationId": "getFunctionDomain",
        "tags": [
          "function-domains"
        ],
        "summary": "Get the zone claim covering a hostname",
        "description": "Requires `platform:function:function:view`.",
        "parameters": [
          {
            "name": "hostname",
            "in": "path",
            "required": true,
            "description": "any hostname covered by a claimed zone, or the zone apex itself — resolves to the zone's one claim",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The domain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "HOSTNAME_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "FunctionDomain_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "releaseFunctionDomain",
        "tags": [
          "function-domains"
        ],
        "summary": "Release (delete) a claimed zone",
        "description": "Requires `platform:function:domain:manage`.",
        "parameters": [
          {
            "name": "hostname",
            "in": "path",
            "required": true,
            "description": "any hostname covered by the claimed zone, or the zone apex itself — resolves to the zone's one claim",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Released"
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "FunctionDomain_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DOMAIN_IN_USE — a public route exists on this zone's apex or on a hostname covered by it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/function-routes": {
      "get": {
        "operationId": "listFunctionRoutes",
        "tags": [
          "function-domains"
        ],
        "summary": "List public routes by hostname or by function address",
        "description": "Requires `platform:function:function:view`.",
        "parameters": [
          {
            "name": "hostname",
            "in": "query",
            "required": false,
            "description": "list every route for this hostname",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "query",
            "required": false,
            "description": "list this one function's routes",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The routes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunctionRouteResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "FUNCTION_ROUTE_FILTER_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Function_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "The platform's canonical error envelope (CONVENTIONS.md §4).",
        "properties": {
          "error": {
            "type": "string",
            "description": "the machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "a human-readable message"
          },
          "details": {
            "type": "object",
            "description": "optional structured detail; omitted when empty",
            "additionalProperties": true
          }
        },
        "required": [
          "error",
          "message"
        ],
        "additionalProperties": false
      },
...
      "ClaimRequest": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          }
        },
        "required": [
          "hostname"
        ],
        "additionalProperties": false
      },
      "DomainRecordView": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TXT"
            ]
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "name",
          "value"
        ],
        "additionalProperties": false
      },
      "DomainVerificationView": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED"
            ]
          },
          "record": {
            "$ref": "#/components/schemas/DomainRecordView"
          }
        },
        "required": [
          "state"
        ],
        "additionalProperties": false
      },
      "DomainResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "verification": {
            "$ref": "#/components/schemas/DomainVerificationView"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "hostname",
          "owner",
          "verification",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "FunctionRouteResponse": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "pathPrefix": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "aliasPrefixes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "hostname",
          "pathPrefix",
          "address",
          "aliasPrefixes"
        ],
        "additionalProperties": false
      },
...
    }
  }
}